Move Root definition to a separate module
